Added connect to service OpenWeatherMap

+++ Models/WeatherForecast.cs
@@ -1,7 +1,7 @@
 public class WeatherForecast
 {
     public string DateFormatted { get; set; }
-    public int TemperatureC { get; set; }
+    public double TemperatureC { get; set; }
     public string Summary { get; set; }
 
     public int TemperatureF
